css-experiments: rename pages
